New app design layout created.
Inside the project, I created a specific controller in order to manage routes for test design purposes. This controller will have to be destroyed in the futur, as pages will have to go to the right controllers. Therefore, forms are not connected to controller or DB : they are juste empty input and back end dev will have to connect with simple form.

I also manage to upgrade rails version and ruby version as the previous versions were outdated and caused the app to crash.