Update in records index UX.
There is now two ways of looking for a record :

1- Search bar.
That would be in 'Rechercher'. Through a search bar, a User can look for a specific reference.

2- Discover
Through a UX that copy cat real shops, a User can dig through all records.

In the futur, it is possible that thos two options could merge in one page. For now I want to test this.